Changing configuration and adding functions to /drops.
Command names are now configurable. Can be provided as a list of aliases
in the config file.

/drops has been modified so kill and killall only work in the world the
player is currently in. Kill now removes items within a defined radius
(default 32 blocks) which is configurable both in the file and at the
time the command is issued. Killall now just removes all items from the
current world, not all worlds.